Corrected a small bug in makeEnclosure().
The bug was found after playing last move of Red in game szkrab6492.
The problem was when an enclosure connected worms so that some of
them could be closed by the opponent in exactly one way (singular), and
some not. Such a case means that by enclosing we cancel some of the opponent
threats.
The bug was not very serious, just singular_dots were too large for one of the threats.