A UT reproducing a bug in getEnclMoves.
See http://eidokropki.reaktywni.pl/#u0O9GbZR
The problem is that when we make a move inside our terr and
we need to enclose to avoid opp's enclosure, then getEnclMoves() sees
both the original terr enclosure and the new, smaller one. The new one
is included in the old one, which is not allowed.