Corrected two bad bugs.
One bug was in sending data back from the worker, where it was shifted
by 4 bytes. As a result, for example, the best move reported by the CNN
was not the proper one, but the one 1 point below the proper one.

The other bug resulted in ignoring config line for the number of workers.
As a result, always 7 was assumed.