Make checkLadder() more universal
Namely, now it can handle situations when the defender
played the dot, or is about to play the dot (to defend).
The function is now oversimplified and does not work
correctly, at least in the new case when the defender
has just played the dot  -- for example, if checks if
attacker dots are connected, but does not check where
(they could be connected from 'the ladder side').