Add new fields to history
Namely, add possibility to store whether we played where
opponent could enclose (i.e., on the border of opp's enclosure),
and to store atari code, i.e., information, which of the four
neighbours of the dot played were in atari before the move.
Although eventually we will probably need to know which
neighbours were in atari before the move, but no longer are
in atari after move -- this would help with ladder detection.